fix : moved tags to gcpVH
