+++ jobs/gcp-discovery/src/main/java/com/tmobile/pacbot/gcp/inventory/vo/CloudDNSVH.java
@@ -1,20 +1,10 @@
 package com.tmobile.pacbot.gcp.inventory.vo;
 
-import java.util.Map;
 
 public class CloudDNSVH extends GCPVH{
     public String dnsSecConfigState;
     public String dnsName;
 
-    private Map<String, String> tags;
-
-    public Map<String, String> getTags() {
-        return tags;
-    }
-
-    public void setTags(Map<String, String> tags) {
-        this.tags = tags;
-    }
 
     public String getDnsName() {
         return dnsName;
